integrate projectId extraction from token
