+++ package.json
@@ -17,6 +17,7 @@
         "amqplib": "^0.5.3",
         "debug": "^4.1.1",
         "dotenv": "^6.2.0",
+        "jsonwebtoken": "^8.5.1",
         "mongodb": "^3.2.7",
         "mongoose": "^5.4.4",
         "winston": "^3.2.1",
