+++ workers/php/index.js
@@ -1,6 +1,7 @@
 const { Worker, ParsingError, DatabaseError } = require('../../lib/worker');
-const db = require('../../lib/db/mongoose-controller');
+const db = require('../../lib/db/controller');
 const path = require('path');
+const { decode } = require('jsonwebtoken');
 
 require('dotenv').config({ path: path.resolve(__dirname, '.', '.env') });
 
@@ -40,6 +41,14 @@ class PhpWorker extends Worker {
         throw new ParsingError('Message parsing error', err);
       }
 
+      let projectId;
+
+      try {
+        projectId = decode(phpError.token);
+      } catch (err) {
+        throw new ParsingError("Can't decode token", err);
+      }
+
       try {
         payload = this.parseData(phpError);
       } catch (err) {
@@ -47,7 +56,7 @@ class PhpWorker extends Worker {
       }
 
       try {
-        await db.saveEvent({ catcherType: this.type, payload });
+        await db.saveEvent(projectId, { catcherType: this.type, payload });
       } catch (err) {
         // @todo Send unprocessed msg back to queue?
         throw new DatabaseError('Saving event to database error', err);
@@ -82,7 +91,7 @@ class PhpWorker extends Worker {
    * @returns {Obejct}
    */
   parseData(obj) {
-    let payload = { };
+    let payload = {};
 
     payload.title = obj['error_description'] || '';
 
@@ -92,7 +101,7 @@ class PhpWorker extends Worker {
     try {
       let timestamp = obj['http_params']['REQUEST_TIME'];
 
-      payload.timestamp = (new Date(timestamp)).getTime();
+      payload.timestamp = new Date(timestamp).getTime();
     } catch (err) {
       throw new ParsingError('Time parsing error', err);
     }
@@ -100,12 +109,12 @@ class PhpWorker extends Worker {
     // Check optional field 'backtrace'
     if (obj['debug_backtrace'] && obj['debug_backtrace'].length) {
       payload.backtrace = [];
-      obj['debug_backtrace'].forEach((item) => {
+      obj['debug_backtrace'].forEach(item => {
         if (item.file && item.line) {
           payload.backtrace.push({
             file: item.file,
             line: item.line,
-            sourceCode: (item.trace && item.trace.length) ? item.trace : []
+            sourceCode: item.trace && item.trace.length ? item.trace : []
           });
 
           if (!item.trace) {
